Do not update osimage attribute when error happens
This patch delay the time to update the osimage attibute into node
object as we should not do it if exception raised before this step.
This patch do not clean the osimage id from node maybe it is useful
for the user to check the image information. One drawback is that
if we hope to destroy the osimage.